requirements.txt
django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
Pillow==10.0.0
python-decouple==3.8
